convert copy assign names tests to vitest
